Implement basic relative wikilink support
Closes #35. This allows hugo-obsidian to intuit absolute paths from unanchored
wikilink paths. Previously, a note at `directory/note.md` would need to be
linked using `[[directory/note]]` in order for the link to be recognized by this
tool. This change allows the shorter `[[note]]` link to be used. This is
accomplished by first walking the directory tree and storing absolute paths for
every path basename, and using this to substitute link names before rendering them.

There's no test infrastructure for this project, but IWOMM